% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSimulatoR.R
\name{DataSimulatoR}
\alias{DataSimulatoR}
\title{Generate Student Data From File}
\usage{
DataSimulatoR(x)
}
\arguments{
\item{x}{A data frame object from the student survey. Number of rows should be equal to number of responses, Number of columns should be equal to 33.}
}
\value{
A series of csv files for students and pdf grading keys
}
\description{
The DataSimulatoR function creates data files (.csv) for students and grading keys (.pdf) based off of student responses to survey found on github.com/LiamOMueller/UCSanDiego_BILD5
}
